feat: use MakeHTTPServer on echo template
