+++ cmd/program/program.go
@@ -21,7 +21,7 @@ type Project struct {
 }
 
 type Framework struct {
-	packageName string
+	packageName []string
 	templater   Templater
 }
 
@@ -31,13 +31,13 @@ type Templater interface {
 	Routes() []byte
 }
 
-const (
-	chiPackage     = "github.com/go-chi/chi/v5"
-	gorillaPackage = "github.com/gorilla/mux"
-	routerPackage  = "github.com/julienschmidt/httprouter"
-	ginPackage     = "github.com/gin-gonic/gin"
-	fiberPackage   = "github.com/gofiber/fiber/v2"
-	echoPackage    = "github.com/labstack/echo/v4"
+var (
+	chiPackage     = []string{"github.com/go-chi/chi/v5"}
+	gorillaPackage = []string{"github.com/gorilla/mux"}
+	routerPackage  = []string{"github.com/julienschmidt/httprouter"}
+	ginPackage     = []string{"github.com/gin-gonic/gin"}
+	fiberPackage   = []string{"github.com/gofiber/fiber/v2"}
+	echoPackage    = []string{"github.com/labstack/echo/v4", "github.com/labstack/echo/v4/middleware"}
 
 	cmdApiPath         = "cmd/api"
 	internalServerPath = "internal/server"
@@ -52,14 +52,13 @@ func (p *Project) ExitCLI(tprogram *tea.Program) {
 }
 
 func (p *Project) createFrameworkMap() {
-
 	p.FrameworkMap["chi"] = Framework{
 		packageName: chiPackage,
 		templater:   tpl.ChiTemplates{},
 	}
 
 	p.FrameworkMap["standard library"] = Framework{
-		packageName: "",
+		packageName: []string{},
 		templater:   tpl.StandardLibTemplate{},
 	}
